Change timer method
The original timer method using setInterval ran into issues if the
callback took too long, as sometimes happened if we had slow commands.
This would cause the callback calls to sometimes overlap each other.
This change uses one-shot timers that don't get set until the callback
returns. It means the interval is lengthened by the time the callback
takes to run, but this is fine as the times aren't intented to be
precise, and it prevents multiple callbacks from running at once.